Move extensions to extension package
