+++ android/app/src/main/kotlin/ca/etsmtl/applets/etsmobile/domain/CheckUserCredentialsValidUseCase.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.Transformations
 import ca.etsmtl.applets.etsmobile.R
 import ca.etsmtl.applets.etsmobile.presentation.App
-import ca.etsmtl.applets.etsmobile.util.isDeviceConnected
+import ca.etsmtl.applets.etsmobile.extension.isDeviceConnected
 import ca.etsmtl.applets.repository.data.model.Resource
 import ca.etsmtl.applets.repository.data.repository.signets.InfoEtudiantRepository
 import model.Etudiant
