+++ android/app/src/main/kotlin/ca/etsmtl/applets/etsmobile/extension/BottomNavigationViewExt.kt
@@ -1,4 +1,4 @@
-package ca.etsmtl.applets.etsmobile.util
+package ca.etsmtl.applets.etsmobile.extension
 
 import android.animation.Animator
 import androidx.core.view.isVisible
