+++ android/app/src/main/kotlin/ca/etsmtl/applets/etsmobile/extension/CircleProgressViewExt.kt
@@ -1,4 +1,4 @@
-package ca.etsmtl.applets.etsmobile.util
+package ca.etsmtl.applets.etsmobile.extension
 
 import android.animation.ArgbEvaluator
 import com.moos.library.CircleProgressView
