+++ android/app/src/main/kotlin/ca/etsmtl/applets/etsmobile/presentation/dashboard/card/grades/GradesCardViewModel.kt
@@ -9,7 +9,7 @@ import ca.etsmtl.applets.etsmobile.domain.FetchCurrentSessionGradesCoursesUseCas
 import ca.etsmtl.applets.etsmobile.presentation.App
 import ca.etsmtl.applets.etsmobile.util.Event
 import ca.etsmtl.applets.etsmobile.util.RefreshableLiveData
-import ca.etsmtl.applets.etsmobile.util.adjustCote
+import ca.etsmtl.applets.etsmobile.extension.adjustCote
 import ca.etsmtl.applets.repository.data.model.Resource
 import model.Cours
 import presentation.ViewModel
