+++ android/app/src/main/kotlin/ca/etsmtl/applets/etsmobile/presentation/gradesdetails/GradesDetailsViewModel.kt
@@ -12,7 +12,7 @@ import ca.etsmtl.applets.etsmobile.domain.FetchGradesDetailsUseCase
 import ca.etsmtl.applets.etsmobile.presentation.App
 import ca.etsmtl.applets.etsmobile.util.Event
 import ca.etsmtl.applets.etsmobile.util.RefreshableLiveData
-import ca.etsmtl.applets.etsmobile.util.toLocalizedString
+import ca.etsmtl.applets.etsmobile.extension.toLocalizedString
 import ca.etsmtl.applets.repository.data.model.Resource
 import ca.etsmtl.applets.repository.util.zeroIfNullOrBlank
 import com.shopify.livedataktx.map
