+++ android/app/src/main/kotlin/ca/etsmtl/applets/etsmobile/presentation/login/LoginViewModel.kt
@@ -14,7 +14,7 @@ import ca.etsmtl.applets.etsmobile.domain.FetchSavedSignetsUserCredentialsUserCa
 import ca.etsmtl.applets.etsmobile.domain.SaveSignetsUserCredentialsUseCase
 import ca.etsmtl.applets.etsmobile.presentation.App
 import ca.etsmtl.applets.etsmobile.util.Event
-import ca.etsmtl.applets.etsmobile.util.call
+import ca.etsmtl.applets.etsmobile.extension.call
 import ca.etsmtl.applets.repository.data.model.Resource
 import model.SignetsUserCredentials
 import model.UniversalCode
