+++ android/app/src/main/kotlin/ca/etsmtl/applets/etsmobile/presentation/profile/ProfileViewModel.kt
@@ -11,7 +11,7 @@ import ca.etsmtl.applets.etsmobile.domain.FetchProgrammesUseCase
 import ca.etsmtl.applets.etsmobile.presentation.App
 import ca.etsmtl.applets.etsmobile.util.Event
 import ca.etsmtl.applets.etsmobile.util.RefreshableLiveData
-import ca.etsmtl.applets.etsmobile.util.getGenericErrorMessage
+import ca.etsmtl.applets.etsmobile.extension.getGenericErrorMessage
 import ca.etsmtl.applets.repository.data.model.Resource
 import ca.etsmtl.applets.repository.util.zipResourceTo
 import com.shopify.livedataktx.map
