+++ android/app/src/test/kotlin/ca/etsmtl/applets/etsmobile/presentation/GradesViewModelTest.kt
@@ -7,7 +7,7 @@ import ca.etsmtl.applets.etsmobile.R
 import ca.etsmtl.applets.etsmobile.domain.FetchGradesCoursesUseCase
 import ca.etsmtl.applets.etsmobile.presentation.grades.GradesViewModel
 import ca.etsmtl.applets.etsmobile.util.Event
-import ca.etsmtl.applets.etsmobile.util.adjustCote
+import ca.etsmtl.applets.etsmobile.extension.adjustCote
 import ca.etsmtl.applets.etsmobile.util.mockNetwork
 import ca.etsmtl.applets.repository.data.model.Resource
 import com.nhaarman.mockito_kotlin.capture
